Adds nullability rules to be mostly consistent with protobuf
Protocol Buffers already offer consistent nullability rules. This change
adds those rules to the generated schema in an effort to provide a better
set of guarantees to the frontend. This should help minimize the amount
of defensive code in the frontend and provide much more consistent nullability
boundaries.

Because GraphQL cannot uphold nullability rules on errors, edges that
resolve arrays _can_ be nil, unlike the empty rules in protobuf

Transitions the project to use go modules because that exists now